Fix: application des fix demandés, unification du bouton renommer/supprimer le bouton renommer de la page calandrier/rétrécir le bouton de renommer
