+++ LLIO2025/package.json
@@ -37,7 +37,6 @@
     "@opentelemetry/resources": "^2.5.0",
     "@opentelemetry/sdk-trace-web": "^2.5.0",
     "@tailwindcss/vite": "^4.1.18",
-    "dompurify": "^3.3.3",
     "felte": "^1.3.0",
     "lucide-svelte": "^0.563.0",
     "svelte-select": "^5.8.3",
