+++ LLIO2025/src/Components/Projects/DashboardPaneProjectItem.svelte
@@ -6,9 +6,6 @@
   import { getHoursColor } from '../../utils/displayUtils';
   import { calculateEmployeeTime, calculateRemainingTime } from '../../utils/CalculUtils';
   import type { Category } from '../../Models';
-  import TextInputModal from '../TextInputModal.svelte';
-  import DOMPurify from 'dompurify';
-  import { CategoryApiService } from '../../services/CategoryApiService';
 
   let { project } = $props();
 
@@ -25,23 +22,6 @@
 
   let selectedCategory: Category = $state();
   let enableRenameCategoryPrompt = $state(false);
-
-  function handleRenameCategory(category: Category) 
-  {
-    selectedCategory = category;
-    enableRenameCategoryPrompt = true;
-  }
-
-  async function sendRenameCategory(category: Category, newName: string) {
-    try {
-      return await CategoryApiService.changeCategoryName(newName, category);
-    }
-    catch (error) {
-      alert("Erreur - impossible de modifier le nom de la catégorie")
-    }
-    return false;
-  }
-
 </script>
 
 <div class="border-l-10 border-b" style="border-left-color: {project.color}">
@@ -92,15 +72,6 @@
           >
             <td class="py-2 pl-4 text-left">
               {category.name}
-              {#if category.name !== "Par défaut" && category.id === hoveredCategoryId}
-                        <button
-                            class="justify-end p-2 text-gray-500 hover:text-gray-700 hover:bg-gray-100 rounded-full transition-colors"
-                            onclick={() => handleRenameCategory(category)}
-                            aria-label="Renommer la catégorie"
-                        >
-                        <Pencil size={10} />
-                    </button>
-                {/if}
             </td>
 
 
@@ -151,21 +122,3 @@
 
   </div>
 </div>
-
-
-{#if enableRenameCategoryPrompt}
-    <TextInputModal
-      modalTitle="Modification du nom de la catégorie"
-      modalText={`Renommez la catégorie "<strong>` + DOMPurify.sanitize(selectedCategory.name) + `</strong>"`}
-      defaultTextInValue={DOMPurify.sanitize(selectedCategory.name)}
-      onSuccess={async (val: string) => {
-        if (await sendRenameCategory(selectedCategory, val)) {
-          categories = categories.map((cat) => cat.id === selectedCategory.id ? {...cat, name: val} : cat);
-        }
-        enableRenameCategoryPrompt = false;
-      }}
-      onClose={() => {
-        enableRenameCategoryPrompt = false;
-      }}
-    />
-{/if}
